LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/text/2d.text.measure.emHeights-zero-descent.worker-expected.txt


FAIL Testing emHeights with zero descent metric assert_equals: ctx.measureText('A').emHeightAscent === 40 (got 30[number], expected 40[number]) expected 40 but got 30
